Support fid which is longer than 2147483647

+++ src/main/java/fm/last/moji/MojiFileAttributes.java
@@ -8,7 +8,7 @@ public interface MojiFileAttributes {
 
   long getLength();
 
-  int getFid();
+  long getFid();
 
   String getDomain();
